requirements.txt
# flask>=3.0.2
# transformers>=4.38.2
# torch>=2.2.1
# accelerate>=0.27.2
# sentencepiece>=0.2.0

Flask==3.1.0
langchain-google-genai==2.1.3
google-genai>=1.12.1
python-dotenv>=1.0.0
